+++ include/utils.h
@@ -55,7 +55,7 @@ namespace diskann {
 
   enum Metric { L2 = 0, INNER_PRODUCT = 1, FAST_L2 = 2, PQ = 3 };
 
-  DISKANN_DLLEXPORT float calc_recall_set(unsigned  num_queries,
+  DISKANN_DLLEXPORT double calc_recall_set(unsigned  num_queries,
                                           unsigned *gold_std, float *gs_dist,
                                           unsigned  dim_gs,
                                           unsigned *our_results,
@@ -71,8 +71,8 @@ namespace diskann {
 
   template<typename T>
   DISKANN_DLLEXPORT int build_merged_vamana_index(
-      std::string base_file, diskann::Metric _compareMetric, size_t L, size_t R,
-      float sampling_rate, double ram_budget, std::string mem_index_path);
+      std::string base_file, diskann::Metric _compareMetric, unsigned L, unsigned R,
+      double sampling_rate, double ram_budget, std::string mem_index_path);
 
   inline void alloc_aligned(void **ptr, size_t size, size_t align) {
     *ptr = nullptr;
